refactor: time format 함수 util 폴더로 분리
